[#97][feature] - added parsing for start and end times from an image - moved parsing logic from tesseract.js to tesseractRepository.js - added documentation
